src/pyproject.toml
[project]
name = "intentc"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "PyYAML>=6.0",
    "typer>=0.9",
    "rich>=13.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["intentc"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[project.scripts]
intentc = "intentc.cli.main:app"
